requirements.txt
flask
openai
psycopg2-binary
python-dotenv
llama-index